fix: update additional test files for consistency and compilation
Update remaining test files to maintain consistency with neurosymbolic
architecture and resolve minor compilation issues:

- api_integration_mocks.rs: Fix import paths and mock configurations
- api_integration_tests.rs: Update component client references
- graph_daa_integration_test.rs: Align message bus field names
- integration_fix_validation.rs: Update neural classifier imports
- london_tdd_integration.rs: Fix dependency references
- week3_integration_tests.rs: Resolve ownership/lifetime issues

These files had ancillary changes to support the main 4 target files
and ensure overall test suite compilation consistency.

+++ tests/api_integration_tests.rs
@@ -4,7 +4,6 @@
 
 use std::time::{Duration, Instant};
 use tokio::time::timeout;
-use uuid::Uuid;
 use serde_json::json;
 
 // Simple structures for testing
@@ -170,7 +169,7 @@ mod simple_api_tests {
         
         for (i, task_result) in results.into_iter().enumerate() {
             let (task_id, duration) = task_result.unwrap();
-            assert_eq!(task_id, i, "Task ID should match");
+            assert_eq!(task_id, i as u64, "Task ID should match");
             assert!(duration.as_millis() < 1000, "Task should complete quickly");
         }
